[refactor] coverLetterContent의 함수 lib로 분리
