+++ frontend/src/features/coverLetter/libs/caret.ts
@@ -1,0 +1,32 @@
+// caret 저장
+export const saveCaret = (el: HTMLElement): number => {
+  const sel = window.getSelection();
+  if (!sel || sel.rangeCount === 0) return 0;
+  const range = sel.getRangeAt(0);
+  const preRange = range.cloneRange();
+  preRange.selectNodeContents(el);
+  preRange.setEnd(range.startContainer, range.startOffset);
+  return preRange.toString().length;
+};
+
+// caret 복원
+export const restoreCaret = (el: HTMLElement, offset: number) => {
+  const sel = window.getSelection();
+  if (!sel) return;
+
+  const walker = document.createTreeWalker(el, NodeFilter.SHOW_TEXT);
+  let remaining = offset;
+
+  while (walker.nextNode()) {
+    const node = walker.currentNode as Text;
+    if (remaining <= node.length) {
+      const range = document.createRange();
+      range.setStart(node, remaining);
+      range.collapse(true);
+      sel.removeAllRanges();
+      sel.addRange(range);
+      return;
+    }
+    remaining -= node.length;
+  }
+};
